Add tool to merge PNG files into a KMZ file with timespans.
Given a folder containing png files like:
folder/
  IMG_2000-01-01.png
  IMG_2000-01-09.png
  IMG_2000-01-17.png
  ...

The tool creates a KML file which set the ground overlay as each PNG files.
Also, set the TimeSpan elements based on the dates parsed from file names.

Then, the tool packs the PNG files and the KML file into a KMZ file.
The KMZ file can be opened in Google Earth to view the animation of PNG
files as time changes.